create rust symplectic product map
